# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Wazin وازِن - AI-powered financial education platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile (language, theme, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /simulation/start:
    post:
      operationId: startSimulation
      tags: [simulation]
      summary: Start or restart simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationStartInput"
      responses:
        "200":
          description: New simulation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationState"
  /simulation/allocate:
    post:
      operationId: allocateBudget
      tags: [simulation]
      summary: Allocate budget across categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetAllocationInput"
      responses:
        "200":
          description: Updated simulation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationState"
...
  /events/{eventId}/decide:
    post:
      operationId: makeDecision
      tags: [events]
      summary: Make a decision on an economic event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "200":
          description: Decision result with XP and AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DecisionResult"
...
  /rewards/{rewardId}/redeem:
    post:
      operationId: redeemReward
      tags: [rewards]
      summary: Redeem a virtual reward for a real offer
      parameters:
        - name: rewardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Redemption confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedemptionResult"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, username, email, xp, level, coins, language, theme, avatarUrl, joinedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        xp:
          type: integer
        level:
          type: integer
        coins:
          type: integer
        language:
          type: string
          enum: [ar, en]
        theme:
          type: string
          enum: [light, dark]
        avatarUrl:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
        joinedAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        language:
          type: string
          enum: [ar, en]
        theme:
          type: string
          enum: [light, dark]
        username:
          type: string
...
    SimulationState:
      type: object
      required: [id, month, year, totalBudget, allocations, healthScore, xpEarned, status]
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        totalBudget:
          type: number
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/BudgetAllocation"
        healthScore:
          type: integer
        xpEarned:
          type: integer
        status:
          type: string
          enum: [active, paused, completed]
        pendingEvents:
          type: integer
    BudgetAllocation:
      type: object
      required: [category, categoryAr, allocated, spent, limit]
      properties:
        category:
          type: string
        categoryAr:
          type: string
        allocated:
          type: number
        spent:
          type: number
        limit:
          type: number
        icon:
          type: string
    SimulationStartInput:
      type: object
      required: [monthlyIncome]
      properties:
        monthlyIncome:
          type: number
        difficulty:
          type: string
          enum: [easy, medium, hard]
    BudgetAllocationInput:
      type: object
      required: [allocations]
      properties:
        allocations:
          type: array
          items:
            type: object
            required: [category, amount]
            properties:
              category:
                type: string
              amount:
                type: number
...
    DecisionInput:
      type: object
      required: [optionId]
      properties:
        optionId:
          type: string
    DecisionResult:
      type: object
      required: [xpEarned, coinsEarned, budgetImpact, feedbackEn, feedbackAr, newHealthScore]
      properties:
        xpEarned:
          type: integer
        coinsEarned:
          type: integer
        budgetImpact:
          type: number
        feedbackEn:
          type: string
        feedbackAr:
          type: string
        newHealthScore:
          type: integer
        leveledUp:
          type: boolean
        newLevel:
          type: ["integer", "null"]
...
    RedemptionResult:
      type: object
      required: [success, redemptionCode, messageEn, messageAr]
      properties:
        success:
          type: boolean
        redemptionCode:
          type: ["string", "null"]
        messageEn:
          type: string
        messageAr:
          type: string
...