# lib/api-zod/src/generated/types/budgetAllocation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Wazin وازِن - AI-powered financial education platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetAllocation {
  category: string;
  categoryAr: string;
  allocated: number;
  spent: number;
  limit: number;
  icon?: string;
}